requirements.txt
llvmlite==0.46.0
numba==0.63.1
numpy==2.2.6
PyOpenGL==3.1.10
PyOpenGL-accelerate==3.1.10
PySDL2==0.9.17
